fix(summary): post a chat marker from the tool path too
The tool path (used by the sidecar post-gen writer) hid its summarized range
but never posted a marker — only the interval/slash-command path did. That left
orphaned hidden messages: removed from the model context, still rendered on
screen, with nothing accounting for them. Observed live: 16 messages had been
hidden this way with no visible trace at all.

It also broke attribution in the collapse UI. The fallback range-finder walks
back over the contiguous hidden run above a marker, so with no marker of their
own those 16 were claimed by the next summary down, which reported "22 messages
this replaced" when it had covered 6.

postSummaryMarker moves to summary-collapse.js so both paths can share it
without an import cycle (summary-runner already imports tools/summarize.js).
The pending-hide record now carries the title and summary text so the deferred
flush, which is where the real range is known, can post the marker.

No change was needed to bound the fallback: the walk already stops at a previous
summary marker. It only over-reached because the marker did not exist.
